# Switch to pointer to policy rule, visit and short circuit during authorization