@deads2k 2017-04-13
I'm generally anti-visitor (I think it adds complication and confusion), but if there's proof it actually solves a measured memory problem, this lgtm.

I'm fine merging, then checking and if it doesn't help, reverting.