@smarterclayton 2017-04-13
Isn't this exactly the same problem we fixed in openshift, to use a visitor, which I asked if we regressed when we ported the full authorizer upstream?