@deads2k 2017-04-13
> Isn't this exactly the same problem we fixed in openshift, to use a visitor, which I asked if we regressed when we ported the full authorizer upstream?

Pretty sure that openshift doesn't visit.  I probably wrote it to short-circuit without visiting.