@smarterclayton 2017-04-13
My change was to use a visitor pattern for RoleBindings, not for Rules.