@sttts 2017-04-18
Some nits.