@wojtek-t 2017-04-18
I've run some experiment using 500-node cluster and, before it was significant percentage of the profile:
https://github.com/kubernetes/kubernetes/issues/40015#issue-201288657
with this change it's not present in top 10 elements. So I would treat it as fix to the issue I reported.