@smarterclayton 2017-04-20
/lgtm

based on previous feedback and personal review